zebra: check multipath before using recursively resolved blackhole nh
In case a static route is resolving over multiple nexthops (multipath),
if the first nexthop in the list is recursively resolving over a
blackhole, zebra does not bother looking at other nexthops and sends a
netlink update to the kernel marking the route as a blackhole. This
results in complete traffic blackholing even in a multipath setup.

If number of nexthops > 1, do not bother checking for recursively
resolved blackhole nexthop. Let zebra's multipath handling take over.

Signed-off-by: Abhishek Naik <bhini@amazon.com>